+++ CHANGELOG.md
@@ -4,11 +4,19 @@
 
 ### Added
 
+* Add support for experimental foreach operator (#96)
 * Throw ContextualRuntimeException in user function visitor 
 
 ### Changed
 
 * Add support for escaped variable names
+
+## 0.1.10-3 - 2018-09-19
+
+### Changed
+
+* Fix a bug in InnerJoinOperation when requested order does not include all the identifiers
+
 
 ## 0.1.10-2 - 2018-09-04